Avoid integer truncation in BigDecimal
In a 64-bit program, if mb is > 2^32, the amount to allocate will be truncated to a 32-bit unsigned integer, but then the amount to memset will use the full 64-bit number."

 - Fixes 7697476



git-svn-id: http://svn.macosforge.org/repository/ruby/MacRuby/trunk@3788 23306eb0-4c56-4727-a40e-e92c0eb68959

+++ ext/bigdecimal/bigdecimal.c
@@ -2020,7 +2020,7 @@ static int gnAlloc=0; /* Memory allocation counter */
 VP_EXPORT void *
 VpMemAlloc(U_LONG mb)
 {
-    void *p = xmalloc((unsigned int)mb);
+    void *p = xmalloc(mb);
     if(!p) {
         VpException(VP_EXCEPTION_MEMORY,"failed to allocate memory",1);
     }
